avoid reuse of slot offsets after group is freed/recreated
as introduced in commit 069f05b7f4c9e7b6352cb898bc28642a5f5a1a7b, the
cycled offset-within-slot for all newly created slots started with
whatever was initially in the memory in the 16-bit field just below
the slot, which was always zero for mmapped groups. this was fine as
long as the group was not freed, but under freeing and recreation of
the group at the same address, the offset was continually reset,
defeating the purpose of this hardening. recent changes to improve
ability to return memory to the system and avoid overly-eagar
allocation have made this problem more obvious, but it was always
present.

individually-mmapped allocations above the hard mmap threshold have
always used a global counter to cycle offset, but this did not apply
to individually-mmapped allocations small enough to have a size class,
for which the logic to do individual mmap takes place in alloc_group.

this commit refactors and extends the global counter logic to pass the
value into enframe, where enframe is able to make a decision to use it
or the slot's previous offset based on the slot-index/reserved-bytes
header at offset -3; it's zero in freshly created slots and nonzero in
a slot that's previously been used.

in order to use the global counter, though, we need a way to reduce
its range to the slack size for the requested allocation size n and
the slot size being used. previously, out-of-range values were folder
to zero, but that would result in an offset of zero most of the time.
when the input (from the counter) is not already in the slack range.
ideally, we would just use modular reduction here, but it ranges from
mildly to very slow depending on arch.

instead, perform a sort of fake modular reduction by reducing mod the
power-of-two ceiling of the slack size (bitwise and operation), then
conditionally-subtracting the slack size if still over.

this also improves the situation when a slot that was previously used
for a smaller size (larger slack) is reused for a larger size (smaller
slack). previously, a cycled offset that exceeded the new slack limit
would cycle to zero; now, it gets reduced via the above-described
procedure.