only skip use of fully-free groups if they're freeable
skipping fully-free groups only makes sense when we're trying to keep
them in a state that will be freeable (unmappable) in the future. for
groups that are not freeable that does not help.

there are presently no non-freeable groups, but integration with musl
will introduce them via gap donation from ldso. in the future they may
arise in other ways too.