refactor conditional skipping of memset in mmap-serviced calloc
logically, the logic for when it's okay to skip memset belongs in the
body of calloc, but it's split out to the file defining malloc for
integration with musl, where calloc also needs to deal with issues
like malloc interposition.

rather than just optimizing the case where the hard mmap threshold is
exceeded, elide memset for all non-reusable single-slot groups. the
logic here is the same as what okay_to_free uses to force unmap.